ksyzrecip: remove multiple files outputting and add trigger functions
Outputting a file per flattened structure was indeed pretty messy, and it
had to be considered separately in code, so for the sake of convenience
it was removed.

Other than that, ksyzrecip now generates trigger functions which should
result in completely working kflat module.

Although there is still a matter of prettifying these recipes a bit, they
should at least compile. As for working properly on device, we will see.