ksyzrecip: refactor
ksyzrecip was previously a bit of a mess.
The code was allover the place, the API was a mess, and now that it
actually works and does what it was supposed to do from the beggining,
is the time to refactor it to a state in which it is a bit more
workable with and more readable.