uflat: introduce error codes
uflat error reporting convention was to log to stdout, which might be
seen as undesirable as a library.

This commit introduces error codes in addition to the logs.
UFLAT error codes were done a bit differently to Unflatten, since UFLAT
integrates more tightly with the base flatten code, which uses errno and
basic UNIX error codes, they were propagated to UFLAT as well for this
reason.

Unfortunately, UFLAT has functions returning pointers that can error out
in various ways, so just returning a NULL won't do. Here two approaches
were considered, use errno style variable or incorporate the error value
into the returned pointer. Out of these two, the latter was chosen, to
not maintain more internal state of an already complex system.

Because the UNIX error codes are not as descriptive for our case, in
opposition to Unflatten in which the status codes were written with
the library in mind, the logging is still there as it provides valuable
information about the error that occurred. Now it is opt in with
UFLAT_LOGGING define.

Also, UFLAT is now compiled as C code, as a result of introducing
ERR_PTR, because of C++'s restrictive type casting rules.